+++ Front-End/src/styles/css/pages/WorkContract.css
@@ -11,7 +11,7 @@
     padding: 10px;
     border: 1px solid #ccc; /* 그리드 아이템에 테두리 설정 */
     grid-template-columns: 1fr 2fr; /* 부모그리드가사용가능한 넓이중 3/1 3/2 만큼 차지*/
-    grid-template-rows: 800px; /*행의 높이 800으로 고정 그런데 grid cell item에 직접 명시하면 그거 따라감.*/
+    grid-template-rows: 580px; /*행의 높이 800으로 고정 그런데 grid cell item에 직접 명시하면 그거 따라감.*/
     /*minmax(790px, 1fr); 최소 크기 ~ 최대크기 1fr  1/1 만큼 차지*/
     /*grid-auto-rows : 150px * 자동으로 열이 추가될 경우 div태그 */
     grid-column-gap: 30px; /* 열 공백 주기 */
@@ -21,7 +21,7 @@
   .wc-grid-cell-item1 {
     background-color: white;
     position: relative;
-    height: 746px;
+    
     border-right: solid 1px lightgray;
     border-left:solid 1px lightgray;
     
@@ -31,6 +31,10 @@
   .wc-grid-cell-item2 {
     background-color: white;
     
+    overflow-y: auto;
+    
+    border-right: 1px solid #ccc;
+    border-bottom: 1px solid #ccc;
   }
   
   .wc-table-scroll {
@@ -81,6 +85,7 @@
   
   .wc-right-head {
     background-color: white;
+    border-top: solid 2px #ffffff;
   }
   
   .wc-right-first-table {
@@ -149,21 +154,17 @@
   
   .wc-right-cell-search-button:hover {
     border-color: rgb(51, 118, 233);
-    border-width: 2px;
+    border-width: 1px;
     cursor: pointer;
     
   } 
   /*주소조회 버튼 */
   
   .wcTab1{
-    border :1px solid white;
+    border :1px solid #707070;
     border-radius: 2px;
     padding: 10px;
-    /* padding-top: 0px;
-    padding-left: 10px;
-    padding-right: 10px;
-    padding-bottom: 10px; */
-    font-size: 17px;
+    font-size: 14px;
     font-weight: bold;
     
   } 
@@ -171,17 +172,46 @@
   
 
   .wcTab2{
-    border :1px solid white;
+    border :1px solid #707070;
     border-radius: 2px;
     padding: 10px;
-    font-size: 17px;
+    font-size: 14px;
     font-weight: bold;
   }
   
   
+/* 작성,조회중 button이 클릭 됐을때 Click했을때만 class에 add 될 것.*/
   .on {
     background-color: #1985fa;
     color: #ffffff;
+    font-size: 14px;
+    
     
   }
-  /* 작성,조회중 button이 클릭 됐을때 Click했을때만 class에 add 될 것.*/
+
+
+  /*Modal style*/
+  .wcModal1 {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.5);
+    z-index: 100px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  .wcModal2 {
+    background-color: white;
+    padding: 30px;
+    border-radius: 8px;
+    box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.2);
+    width: 30%;
+    height: 55%;
+    max-height: 100%;
+    overflow-y: auto;
+  }
+
